Incorporate PR review comments
